+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.6.0-dev
   **BUG FIXES**
   - [#140](https://github.com/Datatamer/unify-client-python/issues/140) Dataset `itergeofeatures` now supports data with geo attribute NULL
+  - [#123](https://github.com/Datatamer/unify-client-python/issues/123) Fix base_path bug for a custom api endpoint
 
 ## 0.5.0
   **NEW FEATURES**
@@ -14,7 +15,7 @@
   - [#114](https://github.com/Datatamer/unify-client-python/issues/114) Add support for generating pairs estimate
   - [#106](https://github.com/Datatamer/unify-client-python/issues/106) Add support for initializing a source dataset
   - [#107](https://github.com/Datatamer/unify-client-python/issues/107) Add support for creating a dataset attribute
-  
+
   **BUG FIXES**
   - [#118](https://github.com/Datatamer/unify-client-python/issues/118) Fix JSON sent for Dataset.update_records